requirements.txt
Pillow==10.0.0
PyQt6>=6.4.0
pandas>=1.5.0
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.0.0
numpy>=1.20.0
natasha>=1.6.0
openpyxl>=3.0.0
nest-asyncio>=1.5.6
bitsandbytes>=0.41.1
python-dotenv>=0.19.0
requests>=2.28.0
python-docx==0.8.11 